add src.services.meetup.events.get_group_events
This commit adds a src.services and a src.services.meetup module.
The latter contains auxiliar classes in it that ultimately help
src.services.meetup.events.get_group_events connect to Meetup and
retrieve the group's events and return the basic information, including
whether the event is recurring or not.