PEP8
Most of the changes here are simple. Some unused imports are being
removed and a really long lines are being wrapped at 79 characters.

There are also a couple of lines that can't really be helped at this
time, so they're being marked with `NOQA`. Hopefully they will be
addressed at some point in the future.