Add initial Vagrantfile
This is the beginning of putting together a Vagrant box for people who
would like to contribute to nycpython.com. The Vagrantfile (via
bootstrap.sh) installs all of the dependencies.

The README is being updated to reflect this change in setup. It's also
being changed from Markdown to reStructuredText. Why? Because Python.

Flask-SQLAlchemy is being removed as a dependency for the time being
because we aren't yet using a database. Once it's placed back,
SQLAlchemy, psycopg2, and PostgreSQL will need to be added as well, with
Postgres being handled by Vagrant.