Correct Vagrant permissions
When using a virtual environment you shouldn't need to install packages
using `sudo`. When Vagrant provisions a box, however, all of its actions
are done as the root user. This means that the virtual environment was
being created by root, therefore requiring all future packages to be
installed by root or a user with root access. Through a combination of
`chown` and `sudo -u`, this requirement is removed.