Fix __all__ for the meetup package
When developing the current version of the `meetup` package, I
originally implemented it as functions, not as a class. When I switched
to the `APIWrapper` class, I neglected to updated `__all__`. This
addresses that.

I'm also taking this opportunity to make a couple of trivia changes that
have no impact on anything.